+++ cmd/compeditor/main.go
@@ -11,6 +11,7 @@ import (
 	"math"
 	"path/filepath"
 	"runtime"
+	"strings"
 	"time"
 
 	glfw "github.com/go-gl/glfw/v3.1/glfw"
@@ -19,6 +20,7 @@ import (
 	gui "github.com/tbogdala/eweygewey"
 	guiinput "github.com/tbogdala/eweygewey/glfwinput"
 	gombz "github.com/tbogdala/gombz"
+	groggy "github.com/tbogdala/groggy"
 
 	fizzle "github.com/tbogdala/fizzle"
 	component "github.com/tbogdala/fizzle/component"
@@ -28,20 +30,32 @@ import (
 )
 
 var (
-	windowWidth         = 1280
-	windowHeight        = 720
-	mainWindow          *glfw.Window
-	camera              *fizzle.OrbitCamera
-	uiman               *gui.Manager
-	renderer            *forward.ForwardRenderer
-	textureMan          *fizzle.TextureManager
-	basicShader         *fizzle.RenderShader
-	basicShaderFilepath = "../../examples/assets/forwardshaders/basicSkinned"
+	windowWidth                = 1280
+	windowHeight               = 720
+	mainWindow                 *glfw.Window
+	camera                     *fizzle.OrbitCamera
+	uiman                      *gui.Manager
+	renderer                   *forward.ForwardRenderer
+	textureMan                 *fizzle.TextureManager
+	colorShader                *fizzle.RenderShader
+	colorShaderFilepath        = "../../examples/assets/forwardshaders/color"
+	basicShader                *fizzle.RenderShader
+	basicShaderFilepath        = "../../examples/assets/forwardshaders/basic"
+	basicSkinnedShader         *fizzle.RenderShader
+	basicSkinnedShaderFilepath = "../../examples/assets/forwardshaders/basicSkinned"
 
 	clearColor = gui.ColorIToV(32, 32, 32, 32)
 
-	visibleMeshes map[string]*componentRenderable
-	theComponent  component.Component
+	shaders      map[string]*fizzle.RenderShader
+	componentMan *component.Manager
+
+	visibleMeshes    map[string]*meshRenderable
+	visibleColliders []*colliderRenderable
+	theComponent     component.Component
+	childComponents  []*component.Component
+
+	// childRefFilenames is a map of child reference filename to component name
+	childRefFilenames map[string]string
 
 	appStartTime time.Time
 	totalTime    float64
@@ -52,7 +66,16 @@ const (
 	fontFilepath = "../../examples/assets/Oswald-Heavy.ttf"
 	fontGlyphs   = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890., :[]{}\\|<>;\"'~`?/-+_=()*&^%$#@!"
 
-	compMeshWindowTitle = "ComponentMesh"
+	compMeshWindowID = "ComponentMesh"
+
+	segsInSphereWire = 32
+
+	// ui layout constants
+	textWidth = 0.2
+	width3Col = 0.8 / 3.0
+	width4Col = 0.2
+	meshWndX  = 0.65
+	meshWndY  = 0.99
 )
 
 // block of flags set on the command line
@@ -61,12 +84,22 @@ var (
 	flagComponentFile string
 )
 
-// componentRenderable is used to tie together state for the component mesh,
+// meshRenderable is used to tie together state for the component mesh,
 // the renderable for this component mesh and any other state information relating.
-type componentRenderable struct {
-	ComponentMesh     *component.ComponentMesh
+type meshRenderable struct {
+	ComponentMesh     *component.Mesh
 	Renderable        *fizzle.Renderable
 	AnimationsEnabled []bool
+}
+
+// colliderRenderable is used to tie together state for the component collider
+// and the renderable for it.
+type colliderRenderable struct {
+	// Collider is a copy of the CollisionRef data used to make the renderable
+	Collider component.CollisionRef
+
+	// Renderable is the drawable wireframe representing the rendreable
+	Renderable *fizzle.Renderable
 }
 
 // GLFW event handling must run on the main OS thread. If this doesn't get
@@ -80,9 +113,9 @@ func init() {
 	flag.StringVar(&flagComponentFile, "cf", "component.json", "the name of the component file to load and save")
 }
 
-// getCompRenderableByName will return a *componentRenderable for a given
+// getMeshRenderableByName will return a *meshRenderable for a given
 // ComponentMesh name from visibleMeshes
-func getCompRenderableByName(compMeshName string) *componentRenderable {
+func getMeshRenderableByName(compMeshName string) *meshRenderable {
 	for _, cr := range visibleMeshes {
 		if cr.ComponentMesh.Name == compMeshName {
 			return cr
@@ -91,7 +124,7 @@ func getCompRenderableByName(compMeshName string) *componentRenderable {
 	return nil
 }
 
-func makeRenderableForMesh(compMesh *component.ComponentMesh) *fizzle.Renderable {
+func makeRenderableForMesh(compMesh *component.Mesh) *fizzle.Renderable {
 	prefixDir := ""
 	if len(flagComponentFile) > 0 {
 		prefixDir, _ = filepath.Split(flagComponentFile)
@@ -129,13 +162,15 @@ func makeRenderableForMesh(compMesh *component.ComponentMesh) *fizzle.Renderable
 		return nil
 	}
 
-	compRenderable := new(componentRenderable)
+	compRenderable := new(meshRenderable)
 	r := fizzle.CreateFromGombz(compMesh.SrcMesh)
-	r.Core.Shader = basicShader
+	r.Core.Shader = basicSkinnedShader
+	r.Location = compMesh.Offset
+	r.Scale = compMesh.Scale
 
 	// Create a quaternion if rotation parameters are set
 	if compMesh.RotationDegrees != 0.0 {
-		r.Rotation = mgl.QuatRotate(mgl.DegToRad(compMesh.RotationDegrees), compMesh.RotationAxis)
+		r.LocalRotation = mgl.QuatRotate(mgl.DegToRad(compMesh.RotationDegrees), compMesh.RotationAxis)
 	}
 
 	// store the new renderable with the component mesh it belongs to
@@ -153,39 +188,16 @@ func makeRenderableForMesh(compMesh *component.ComponentMesh) *fizzle.Renderable
 	return r
 }
 
-func createMeshWindow(newCompMesh *component.ComponentMesh, screenX, screenY float32) {
+func createMeshWindow(newCompMesh *component.Mesh, screenX, screenY float32) {
 	// FIXME: find a better spot to spawn potentially
-	meshWnd := uiman.NewWindow(compMeshWindowTitle, screenX, screenY, 0.35, 0.5, func(wnd *gui.Window) {
-		const textWidth = 0.2
-		const width3Col = 0.8 / 3.0
-		const width4Col = 0.2
-
-		compRenderable := getCompRenderableByName(newCompMesh.Name)
-		if newCompMesh.SrcMesh != nil && compRenderable != nil && len(newCompMesh.SrcMesh.Animations) > 0 {
-			ani := newCompMesh.SrcMesh.Animations[0]
-			wnd.RequestItemWidthMin(textWidth)
-			wnd.Text("Animation: ")
-			wnd.Text(ani.Name)
-
-			wnd.StartRow()
-			wnd.Text(fmt.Sprintf("Duration: %v", ani.Duration))
-			wnd.StartRow()
-			wnd.Text(fmt.Sprintf("Ticks Per Second: %v", ani.TicksPerSecond))
-
-			wnd.StartRow()
-			wnd.Checkbox("RunAnimations", &compRenderable.AnimationsEnabled[0])
-			wnd.Text("Run Animation")
-			if compRenderable.AnimationsEnabled[0] {
-				aniTime := float32(math.Mod(totalTime*float64(ani.TicksPerSecond), float64(ani.Duration)))
-				compRenderable.Renderable.Core.Skeleton.Animate(&ani, aniTime)
-			}
-
-			wnd.Separator()
-		}
-
+	meshWnd := uiman.NewWindow(compMeshWindowID, screenX, screenY, 0.30, 0.75, func(wnd *gui.Window) {
+		compRenderable := getMeshRenderableByName(newCompMesh.Name)
 		wnd.RequestItemWidthMin(textWidth)
 		wnd.Text("Name")
 		wnd.Editbox("meshNameEditbox", &newCompMesh.Name)
+
+		// force the window id to be the mesh name plus a prefix
+		wnd.ID = fmt.Sprintf("%s%s", compMeshWindowID, newCompMesh.Name)
 
 		wnd.StartRow()
 		wnd.RequestItemWidthMin(textWidth)
@@ -341,6 +353,28 @@ func createMeshWindow(newCompMesh *component.ComponentMesh, screenX, screenY flo
 		wnd.Space(textWidth)
 		wnd.Checkbox("MaterialGenerateMips", &newCompMesh.Material.GenerateMipmaps)
 		wnd.Text("Generate Mipmaps")
+
+		// do the user interface for animations
+		if newCompMesh.SrcMesh != nil && compRenderable != nil && len(newCompMesh.SrcMesh.Animations) > 0 {
+			for aniIndex, animation := range newCompMesh.SrcMesh.Animations {
+				if aniIndex == 0 {
+					wnd.Separator()
+					wnd.RequestItemWidthMin(textWidth)
+					wnd.Text("Animation: ")
+				} else {
+					wnd.StartRow()
+					wnd.Space(textWidth)
+				}
+				wnd.Checkbox(fmt.Sprintf("RunAnimations %d", aniIndex), &compRenderable.AnimationsEnabled[0])
+				wnd.Text(animation.Name)
+				if compRenderable.AnimationsEnabled[0] {
+					aniTime := float32(math.Mod(totalTime*float64(animation.TicksPerSecond), float64(animation.Duration)))
+					compRenderable.Renderable.Core.Skeleton.Animate(&animation, aniTime)
+				}
+			}
+
+			wnd.Separator()
+		}
 	})
 	meshWnd.Title = "Mesh Properties"
 	meshWnd.AutoAdjustHeight = false
@@ -348,25 +382,30 @@ func createMeshWindow(newCompMesh *component.ComponentMesh, screenX, screenY flo
 	meshWnd.ShowScrollBar = true
 }
 
-func doLoadComponentFile(filepath string) {
-	existingCompJSON, err := ioutil.ReadFile(filepath)
+func doLoadComponentFile(componentFilepath string) {
+	existingCompJSON, err := ioutil.ReadFile(componentFilepath)
 	if err == nil {
 		err := json.Unmarshal(existingCompJSON, &theComponent)
 		if err != nil {
-			fmt.Printf("Failed to load component %s: %v\n", filepath, err)
+			fmt.Printf("Failed to load component %s: %v\n", componentFilepath, err)
 		} else {
-			fmt.Printf("Loaded component: %s\n", filepath)
+			fmt.Printf("Loaded component: %s\n", componentFilepath)
 
 			// destroy all existing renderables
 			for _, r := range visibleMeshes {
 				r.Renderable.Destroy()
 				r.Renderable = nil
 			}
-			visibleMeshes = make(map[string]*componentRenderable)
+			for _, vc := range visibleColliders {
+				vc.Renderable.Destroy()
+				vc.Renderable = nil
+			}
+			visibleMeshes = make(map[string]*meshRenderable)
+			visibleColliders = make([]*colliderRenderable, 0)
 
 			// open windows for all existing meshes
-			screenX := float32(0.64)
-			screenY := float32(0.70)
+			screenX := float32(meshWndX)
+			screenY := float32(meshWndY)
 			for _, compMesh := range theComponent.Meshes {
 				createMeshWindow(compMesh, screenX, screenY)
 				if compMesh.SrcFile != "" {
@@ -374,6 +413,22 @@ func doLoadComponentFile(filepath string) {
 				}
 				screenX += 0.05
 				screenY -= 0.05
+
+				// load any referenced textures
+				for _, texFile := range compMesh.Material.Textures {
+					prefixDir := ""
+					if len(flagComponentFile) > 0 {
+						prefixDir, _ = filepath.Split(flagComponentFile)
+					}
+
+					texFilepath := prefixDir + texFile
+					_, err := textureMan.LoadTexture(texFile, texFilepath)
+					if err != nil {
+						fmt.Printf("Failed to load texture %s: %v\n", texFile, err)
+					} else {
+						fmt.Printf("Loaded texture: %s\n", texFile)
+					}
+				}
 			}
 		}
 	}
@@ -382,6 +437,11 @@ func doLoadComponentFile(filepath string) {
 func main() {
 	// parse the command line options
 	flag.Parse()
+
+	// Setup the log handlers
+	groggy.Register("INFO", groggy.DefaultSyncHandler)
+	groggy.Register("ERROR", groggy.DefaultSyncHandler)
+	groggy.Register("DEBUG", groggy.DefaultSyncHandler)
 
 	// start off by initializing the GL and GLFW libraries and creating a window.
 	w, gfx := initGraphics("Component Editor", windowWidth, windowHeight)
@@ -412,9 +472,30 @@ func main() {
 	// load the basic shader
 	basicShader, err = fizzle.LoadShaderProgramFromFiles(basicShaderFilepath, nil)
 	if err != nil {
-		panic("Failed to compile and link the color shader program! " + err.Error())
+		panic("Failed to compile and link the basic shader program! " + err.Error())
 	}
 	defer basicShader.Destroy()
+
+	// load the basic skinned shader
+	basicSkinnedShader, err = fizzle.LoadShaderProgramFromFiles(basicSkinnedShaderFilepath, nil)
+	if err != nil {
+		panic("Failed to compile and link the basic skinned shader program! " + err.Error())
+	}
+	defer basicSkinnedShader.Destroy()
+
+	// load the color shader
+	colorShader, err = fizzle.LoadShaderProgramFromFiles(colorShaderFilepath, nil)
+	if err != nil {
+		panic("Failed to compile and link the color shader program! " + err.Error())
+	}
+	defer colorShader.Destroy()
+
+	// setup the component manager
+	shaders = make(map[string]*fizzle.RenderShader)
+	shaders["Basic"] = basicShader
+	shaders["BasicSkinned"] = basicSkinnedShader
+	shaders["Color"] = colorShader
+	componentMan = component.NewManager(textureMan, shaders)
 
 	// setup the camera to look at the component
 	camera = fizzle.NewOrbitCamera(mgl.Vec3{0, 0, 0}, math.Pi/2.0, 5.0, math.Pi/2.0)
@@ -428,13 +509,15 @@ func main() {
 
 	/////////////////////////////////////////////////////////////////////////////
 	// setup the component and user interface
-	visibleMeshes = make(map[string]*componentRenderable)
+	visibleMeshes = make(map[string]*meshRenderable)
+	visibleColliders = make([]*colliderRenderable, 0)
+	childRefFilenames = make(map[string]string)
 
 	// if the component file passed in as a flag exists, try to load it
 	doLoadComponentFile(flagComponentFile)
 
 	// create a window for operating on the component file
-	componentWindow := uiman.NewWindow("Component", 0.74, 0.99, 0.25, 0.25, func(wnd *gui.Window) {
+	componentWindow := uiman.NewWindow("Component", 0.01, 0.99, 0.25, 0.5, func(wnd *gui.Window) {
 		loadComponent, _ := wnd.Button("componentFileLoadButton", "Load")
 		saveComponent, _ := wnd.Button("componentFileSaveButton", "Save")
 		wnd.Editbox("componentFileEditbox", &flagComponentFile)
@@ -453,7 +536,7 @@ func main() {
 		} else if loadComponent {
 			// remove all existing mesh windows
 			meshWindows := uiman.GetWindowsByFilter(func(w *gui.Window) bool {
-				if w.ID == compMeshWindowTitle {
+				if strings.HasPrefix(w.ID, compMeshWindowID) {
 					return true
 				}
 				return false
@@ -468,22 +551,289 @@ func main() {
 		}
 
 		wnd.Separator()
+		wnd.RequestItemWidthMin(textWidth)
 		wnd.Text("Name")
 		wnd.Editbox("componentNameEditbox", &theComponent.Name)
 
-		wnd.StartRow()
-		addMesh, _ := wnd.Button("componentFileAddMeshButton", "+Mesh")
+		// do the user interface for mesh windows
+		wnd.Separator()
+		wnd.RequestItemWidthMin(textWidth)
+		wnd.Text("Meshes:")
+		addMesh, _ := wnd.Button("componentFileAddMeshButton", "Add Mesh")
 		if addMesh {
-			newCompMesh := component.NewComponentMesh()
+			newCompMesh := component.NewMesh()
 			newCompMesh.Name = fmt.Sprintf("Mesh %d", len(theComponent.Meshes)+1)
 			theComponent.Meshes = append(theComponent.Meshes, newCompMesh)
-			createMeshWindow(newCompMesh, 0.64, 0.70)
+			createMeshWindow(newCompMesh, meshWndX, meshWndY)
 		}
+
+		meshesThatSurvive := theComponent.Meshes[:0]
+		for compMeshIndex, compMesh := range theComponent.Meshes {
+			wnd.StartRow()
+			wnd.RequestItemWidthMin(textWidth)
+			wnd.Text(fmt.Sprintf("%s", compMesh.Name))
+			showMeshWnd, _ := wnd.Button(fmt.Sprintf("buttonShowMesh%d", compMeshIndex), "Show")
+			hideMeshWnd, _ := wnd.Button(fmt.Sprintf("buttonHideMesh%d", compMeshIndex), "Hide")
+			deleteMesh, _ := wnd.Button(fmt.Sprintf("buttonDeleteMesh%d", compMeshIndex), "Delete")
+			if showMeshWnd {
+				meshWindow := uiman.GetWindow(fmt.Sprintf("%s%s", compMeshWindowID, compMesh.Name))
+				if meshWindow == nil {
+					createMeshWindow(compMesh, meshWndX, meshWndY)
+				}
+			}
+			if hideMeshWnd || deleteMesh {
+				meshWindow := uiman.GetWindow(fmt.Sprintf("%s%s", compMeshWindowID, compMesh.Name))
+				if meshWindow != nil {
+					uiman.RemoveWindow(meshWindow)
+				}
+			}
+			if !deleteMesh {
+				meshesThatSurvive = append(meshesThatSurvive, compMesh)
+			}
+		}
+		// FIXME: not Destroying renderables for meshes that don't survive
+		theComponent.Meshes = meshesThatSurvive
+
+		// do the user interface for colliders
+		wnd.Separator()
+		wnd.RequestItemWidthMin(textWidth)
+		wnd.Text("Colliders: ")
+		addNewCollider, _ := wnd.Button("buttonAddCollider", "Add Collider")
+		if addNewCollider {
+			// create a new sphere collider
+			newCollider := new(component.CollisionRef)
+			newCollider.Type = component.ColliderTypeSphere
+			newCollider.Radius = 1.0
+			theComponent.Collisions = append(theComponent.Collisions, newCollider)
+		}
+
+		collidersThatSurvive := theComponent.Collisions[:0]
+		visibleCollidersThatSurvive := visibleColliders[:0]
+		for colliderIndex, collider := range theComponent.Collisions {
+			wnd.StartRow()
+			wnd.RequestItemWidthMin(textWidth)
+			wnd.Text(fmt.Sprintf("Collider %d:", colliderIndex))
+
+			delCollider, _ := wnd.Button(fmt.Sprintf("buttonDeleteCollider%d", colliderIndex), "X")
+			prevColliderType, _ := wnd.Button(fmt.Sprintf("buttonPrevColliderType%d", colliderIndex), "<")
+			nextColliderType, _ := wnd.Button(fmt.Sprintf("buttonNextColliderType%d", colliderIndex), ">")
+
+			if !delCollider {
+				collidersThatSurvive = append(collidersThatSurvive, collider)
+
+				if prevColliderType {
+					collider.Type = collider.Type - 1
+					if collider.Type < 0 {
+						collider.Type = component.ColliderTypeCount - 1
+					}
+				}
+				if nextColliderType {
+					collider.Type = collider.Type + 1
+					if collider.Type >= component.ColliderTypeCount {
+						collider.Type = 0
+					}
+				}
+
+				switch collider.Type {
+				case component.ColliderTypeAABB:
+					wnd.Text("Axis Aligned Bounding Box")
+					wnd.StartRow()
+					wnd.Space(textWidth)
+					wnd.RequestItemWidthMin(width4Col)
+					wnd.Text("Min")
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderMinX%d", colliderIndex), 0.01, &collider.Min[0])
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderMinY%d", colliderIndex), 0.01, &collider.Min[1])
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderMinZ%d", colliderIndex), 0.01, &collider.Min[2])
+
+					wnd.StartRow()
+					wnd.Space(textWidth)
+					wnd.RequestItemWidthMin(width4Col)
+					wnd.Text("Max")
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderMaxX%d", colliderIndex), 0.01, &collider.Max[0])
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderMaxY%d", colliderIndex), 0.01, &collider.Max[1])
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderMaxZ%d", colliderIndex), 0.01, &collider.Max[2])
+
+				case component.ColliderTypeSphere:
+					wnd.Text("Sphere")
+					wnd.StartRow()
+					wnd.Space(textWidth)
+					wnd.RequestItemWidthMin(width4Col)
+					wnd.Text("Offset")
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderOffsetX%d", colliderIndex), 0.01, &collider.Offset[0])
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderOffsetY%d", colliderIndex), 0.01, &collider.Offset[1])
+					wnd.RequestItemWidthMax(width4Col)
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderOffsetZ%d", colliderIndex), 0.01, &collider.Offset[2])
+
+					wnd.StartRow()
+					wnd.Space(textWidth)
+					wnd.RequestItemWidthMin(width4Col)
+					wnd.Text("Radius")
+					wnd.DragSliderFloat(fmt.Sprintf("ColliderRadius%d", colliderIndex), 0.01, &collider.Radius)
+				default:
+					wnd.Text(fmt.Sprintf("Unknown collider (%d)!", collider.Type))
+				}
+
+				// see if we need to update the renderable if it exists already
+				if len(visibleColliders) > colliderIndex {
+					visCollider := visibleColliders[colliderIndex]
+
+					switch collider.Type {
+					case component.ColliderTypeAABB:
+						if !visCollider.Collider.Min.ApproxEqual(collider.Min) ||
+							!visCollider.Collider.Max.ApproxEqual(collider.Max) ||
+							visCollider.Collider.Type != collider.Type {
+							visCollider.Collider = *collider
+							visCollider.Renderable = fizzle.CreateWireframeCube(collider.Min[0], collider.Min[1], collider.Min[2],
+								collider.Max[0], collider.Max[1], collider.Max[2])
+						}
+					case component.ColliderTypeSphere:
+						if !visCollider.Collider.Offset.ApproxEqual(collider.Offset) ||
+							math.Abs(float64(visCollider.Collider.Radius-collider.Radius)) > 0.01 ||
+							visCollider.Collider.Type != collider.Type {
+							visCollider.Collider = *collider
+							visCollider.Renderable = fizzle.CreateWireframeCircle(
+								collider.Offset[0], collider.Offset[1], collider.Offset[2], collider.Radius, segsInSphereWire, fizzle.X|fizzle.Y)
+							visCollider.Renderable.AddChild(fizzle.CreateWireframeCircle(
+								collider.Offset[0], collider.Offset[1], collider.Offset[2], collider.Radius, segsInSphereWire, fizzle.Y|fizzle.Z))
+							visCollider.Renderable.AddChild(fizzle.CreateWireframeCircle(
+								collider.Offset[0], collider.Offset[1], collider.Offset[2], collider.Radius, segsInSphereWire, fizzle.X|fizzle.Z))
+						}
+					}
+					// FIXME: not Destroying renderables for visCol's that don't survive
+					visibleCollidersThatSurvive = append(visibleCollidersThatSurvive, visCollider)
+				} else {
+					// append a new visible collider
+					visCollider := new(colliderRenderable)
+					visCollider.Collider = *collider
+
+					switch collider.Type {
+					case component.ColliderTypeAABB:
+						visCollider.Renderable = fizzle.CreateWireframeCube(collider.Min[0], collider.Min[1], collider.Min[2],
+							collider.Max[0], collider.Max[1], collider.Max[2])
+					case component.ColliderTypeSphere:
+						visCollider.Renderable = fizzle.CreateWireframeCircle(
+							collider.Offset[0], collider.Offset[1], collider.Offset[2], collider.Radius, segsInSphereWire, fizzle.X|fizzle.Y)
+						visCollider.Renderable.AddChild(fizzle.CreateWireframeCircle(
+							collider.Offset[0], collider.Offset[1], collider.Offset[2], collider.Radius, segsInSphereWire, fizzle.Y|fizzle.Z))
+						visCollider.Renderable.AddChild(fizzle.CreateWireframeCircle(
+							collider.Offset[0], collider.Offset[1], collider.Offset[2], collider.Radius, segsInSphereWire, fizzle.X|fizzle.Z))
+					}
+					visibleCollidersThatSurvive = append(visibleCollidersThatSurvive, visCollider)
+				}
+			}
+		}
+		theComponent.Collisions = collidersThatSurvive
+		visibleColliders = visibleCollidersThatSurvive
+
+		wnd.Separator()
+		wnd.RequestItemWidthMin(textWidth)
+		wnd.Text("Child Components:")
+		addChildComponent, _ := wnd.Button("addChildComponent", "Add Child")
+		if addChildComponent {
+			newChildRef := new(component.ChildRef)
+			newChildRef.Scale = mgl.Vec3{1, 1, 1}
+			theComponent.ChildReferences = append(theComponent.ChildReferences, newChildRef)
+		}
+
+		childRefsThatSurvive := theComponent.ChildReferences[:0]
+		for childRefIndex, childRef := range theComponent.ChildReferences {
+			wnd.StartRow()
+			wnd.RequestItemWidthMin(textWidth)
+			wnd.Text("File:")
+			removeReference, _ := wnd.Button(fmt.Sprintf("childRefRemove%d", childRefIndex), "X")
+			loadChildReference, _ := wnd.Button(fmt.Sprintf("childRefLoad%d", childRefIndex), "L")
+			wnd.Editbox(fmt.Sprintf("childRefFileEditbox%d", childRefIndex), &childRef.File)
+
+			wnd.StartRow()
+			wnd.Space(textWidth)
+			wnd.RequestItemWidthMin(width4Col)
+			wnd.Text("Offset")
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefLocationX%d", childRefIndex), 0.01, &childRef.Location[0])
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefLocationY%d", childRefIndex), 0.01, &childRef.Location[1])
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefLocationZ%d", childRefIndex), 0.01, &childRef.Location[2])
+
+			wnd.StartRow()
+			wnd.Space(textWidth)
+			wnd.RequestItemWidthMin(width4Col)
+			wnd.Text("Scale")
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefScaleX%d", childRefIndex), 0.01, &childRef.Scale[0])
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefScaleY%d", childRefIndex), 0.01, &childRef.Scale[1])
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefScaleZ%d", childRefIndex), 0.01, &childRef.Scale[2])
+
+			wnd.StartRow()
+			wnd.Space(textWidth)
+			wnd.RequestItemWidthMin(width4Col)
+			wnd.Text("Rot Axis")
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefRotAxisX%d", childRefIndex), 0.01, &childRef.RotationAxis[0])
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefRotAxisY%d", childRefIndex), 0.01, &childRef.RotationAxis[1])
+			wnd.RequestItemWidthMax(width4Col)
+			wnd.DragSliderFloat(fmt.Sprintf("childRefRotAxisZ%d", childRefIndex), 0.01, &childRef.RotationAxis[2])
+
+			wnd.StartRow()
+			wnd.Space(textWidth)
+			wnd.RequestItemWidthMin(width4Col)
+			wnd.Text("Rot Deg")
+			wnd.DragSliderFloat(fmt.Sprintf("childRefRotDeg%d", childRefIndex), 0.1, &childRef.RotationDegrees)
+
+			if !removeReference {
+				childRefsThatSurvive = append(childRefsThatSurvive, childRef)
+			}
+			if loadChildReference {
+				prefixDir := ""
+				if len(flagComponentFile) > 0 {
+					prefixDir, _ = filepath.Split(flagComponentFile)
+				}
+				newChildComponent, err := componentMan.LoadComponentFromFile(prefixDir+childRef.File, childRef.File)
+				if err != nil {
+					fmt.Printf("Failed to load child component: %s\n%v\n", childRef.File, err)
+				} else {
+					fmt.Printf("Loaded child component: %s\n", childRef.File)
+					childComponents = append(childComponents, newChildComponent)
+					childRefFilenames[childRef.File] = newChildComponent.Name
+				}
+			}
+		}
+		theComponent.ChildReferences = childRefsThatSurvive
+
+		// remove any visible child components that no longer have a reference
+		childComponentsThatSurvive := childComponents[:0]
+		for _, ref := range theComponent.ChildReferences {
+			compNameToFind, okay := childRefFilenames[ref.File]
+			if !okay {
+				continue
+			}
+
+			for _, childCompToTest := range childComponents {
+				if compNameToFind == childCompToTest.Name {
+					childComponentsThatSurvive = append(childComponentsThatSurvive, childCompToTest)
+					break
+				}
+			}
+		}
+		childComponents = childComponentsThatSurvive
+
 	})
 	componentWindow.Title = "Component File"
 	componentWindow.ShowTitleBar = false
+	componentWindow.ShowScrollBar = true
+	componentWindow.IsScrollable = true
 	componentWindow.IsMoveable = true
-	componentWindow.AutoAdjustHeight = true
 
 	/////////////////////////////////////////////////////////////////////////////
 	// loop until something told the mainWindow that it should close
@@ -492,6 +842,7 @@ func main() {
 	gfx.Enable(graphics.DEPTH_TEST)
 	gfx.Enable(graphics.PROGRAM_POINT_SIZE)
 	gfx.Enable(graphics.BLEND)
+	gfx.Enable(graphics.LINES)
 	gfx.BlendFunc(graphics.SRC_ALPHA, graphics.ONE_MINUS_SRC_ALPHA)
 
 	lastFrame := time.Now()
@@ -540,6 +891,19 @@ func main() {
 			// draw the thing
 			renderer.DrawRenderable(compRenderable.Renderable, nil, perspective, view, camera)
 		}
+
+		// draw the child components
+		for _, childComponent := range childComponents {
+			r := childComponent.GetRenderable(textureMan, shaders)
+			renderer.DrawRenderable(r, nil, perspective, view, camera)
+		}
+
+		// draw all of the colliders
+		gfx.Disable(graphics.DEPTH_TEST)
+		for _, visCollider := range visibleColliders {
+			renderer.DrawLines(visCollider.Renderable, colorShader, nil, perspective, view, camera)
+		}
+		gfx.Enable(graphics.DEPTH_TEST)
 
 		// draw the user interface
 		uiman.Construct(frameDelta)